src/requirements.txt
paho-mqtt==2.1.0
requests==2.31.0
urllib3==1.26.0
chess==1.10.0
matplotlib==3.8.4
numpy==1.26.4
Pillow==11.1.0
Requests==2.31.0